add check for too long input sent to server
