+++ include/Server.hpp
@@ -84,7 +84,6 @@ class Server
 		void		handleSendError(int fd, const std::string& nick);
 		bool		handleUserInput(int fd);
 		std::vector<std::string>	extractMessagesFromBuffer(User* user);
-		void		broadcastMessage(int senderFd, const std::string& nick, const std::string& message); // TESTIN ONLY
 		void		handleDisconnection(int fd, const std::string& reason, const std::string& source);
 };
