+++ src/CommandRegistration.cpp
@@ -6,13 +6,13 @@
 
 // Handles the `NICK` command for a user. Also part of the initial client registration.
 // Command: `NICK <nickname>`
-bool	Command::handleNick(Server* server, User* user, const std::vector<std::string>& tokens)
+void	Command::handleNick(Server* server, User* user, const std::vector<std::string>& tokens)
 {
 	if (tokens.size() < 2)
 	{
 		logUserAction(user->getNickname(), user->getFd(), "sent NICK without a nickname");
 		user->replyError(431, "", "No nickname given");
-		return false;
+		return;
 	}
 
 	const std::string&	nick = tokens[1];
@@ -23,7 +23,7 @@ bool	Command::handleNick(Server* server, User* user, const std::vector<std::stri
 		logUserAction(user->getNickname(), user->getFd(), std::string("tried to set an invalid nickname: ")
 			+ RED + nick + RESET);
 		user->replyError(432, nick, "Erroneous nickname");
-		return false;
+		return;
 	}
 
 	// Nickname is already in use?
@@ -32,24 +32,22 @@ bool	Command::handleNick(Server* server, User* user, const std::vector<std::stri
 		logUserAction(user->getNickname(), user->getFd(),
 			std::string("tried to set a nickname already in use: ") + YELLOW + nick + RESET);
 		user->replyError(433, nick, "Nickname is already in use");
-		return false;
+		return;
 	}
 
 	user->setNickname(nick);
 	user->tryRegister();
-
-	return true;
 }
 
 // Handles the `USER` command for a user. Also part of the initial client registration.
-// Command: `USER <username> <hostname> <servername> :<realname>`
-bool	Command::handleUser(User* user, const std::vector<std::string>& tokens)
+// Command: `USER <username> <hostname> <servername> <realname>`
+void	Command::handleUser(User* user, const std::vector<std::string>& tokens)
 {
 	if (user->isRegistered())
 	{
 		logUserAction(user->getNickname(), user->getFd(), "tried to resend USER after registration");
 		user->replyError(462, "", "You may not reregister");
-		return false;
+		return;
 	}
 
 	// Enough arguments?
@@ -57,7 +55,7 @@ bool	Command::handleUser(User* user, const std::vector<std::string>& tokens)
 	{
 		logUserAction(user->getNickname(), user->getFd(), "sent invalid USER command (too few arguments)");
 		user->replyError(461, "USER", "Not enough parameters");
-		return false;
+		return;
 	}
 	
 	logUserAction(user->getNickname(), user->getFd(), "sent valid USER command");
@@ -71,26 +69,24 @@ bool	Command::handleUser(User* user, const std::vector<std::string>& tokens)
 	user->setRealname(realname);
 
 	user->tryRegister();
-
-	return true;
 }
 
 // Handles the `PASS` command for a user. This is used to authenticate the user with the server.
 // Command: `PASS <password>`
-bool	Command::handlePass(Server* server, User* user, const std::vector<std::string>& tokens)
+void	Command::handlePass(Server* server, User* user, const std::vector<std::string>& tokens)
 {
 	if (user->isRegistered())
 	{
 		logUserAction(user->getNickname(), user->getFd(), "tried to resend PASS after registration");
 		user->replyError(462, "", "You may not reregister");
-		return false;
+		return;
 	}
 
 	if (tokens.size() < 2)
 	{
 		logUserAction(user->getNickname(), user->getFd(), "sent invalid PASS command (missing password)");
 		user->replyError(461, "PASS", "Not enough parameters");
-		return false;
+		return;
 	}
 
 	// Validate the provided password
@@ -101,12 +97,10 @@ bool	Command::handlePass(Server* server, User* user, const std::vector<std::stri
 	{
 		logUserAction(user->getNickname(), user->getFd(), "provided incorrect password");
 		user->replyError(464, "", "Password incorrect");
-		return false;
+		return;
 	}
 
 	logUserAction(user->getNickname(), user->getFd(), "sent valid PASS command");
 	user->setHasPassed(true);
 	user->tryRegister();
-	
-	return true;
 }
